Firm: reintroduced produce/produceAny/etc.
These are now protected methods called internally by the public supply*
methods.

Moved the stock-checking code from PriceFirm back to Firm and split the
logic into a call to the appropriate produce* method only when needed
(i.e. not in stock).

Firm now provides default methods for supply* and produce+ that should
work most of the time: supply checks stock, uses it if sufficient.
Otherwise it calls the appropriate *produce* method, which do nothing by
default.  produce() is still abstract, because it still needs to
distinguish between mismatch and constraint errors (and because Firm
should be an abstract class because it doesn't actually do anything by
itself).