doxygen: layout; v1.8.2 for c++11; require dot
At least 1.8.2 is required for c++11 support (otherwise things like a
final class won't be recognized and won't show up at all).

Added a DoxygenLayout.xml file generation, to move the class diagrams to
the bottom of the page.

Checked to make sure that dot was found when checking for doxygen in
cmake