Market: doxygen; added quantity(price); buy(B)
Documentation with doxygen.

Added quantity(price) method which returns the quantity that can be
purchased for a given price (essentially the inverse of the existing
price() method).

Added buy(Bundle &assets), which buys as much as possible from the
market using the passed-in assets, reducing the assets and adding the
purchased goods to the assets.