Simulation: Renamed addAgent -> createAgent
The current addAgent (and addGood, addMarket) methods are allowing
unintended and unexpected bad habits such as:

    SomeAgent a(arg1, arg2);
    sim->addAgent(a);
    // ... do something else with a

which really should be:

    SomeAgent a(arg1, arg2);
    auto agent = sim->addAgent(a);
    // ... do something with agent, AND FORGET ABOUT a

because the former is bad: addAgent() makes a *copy* of a, so trying to
use a is not accessing the agent stored in the simulation.

The new methods (createAgent/createGood/createMarket) look like this:

    sim->createAgent<SomeAgent>(arg1, arg2);

so createAgent does the actual construction, passing any arguments to
the templated class constructor.  This makes it considerably clearer,
and is quite similar to the *intended* old mechanism:

   sim->addAgent(SomeAgent(arg1, arg2));

Additionally this avoids a double construction (once before adding, and
then a copy constructor in addAgent).

Other changes:
- Added doxygen docs
- Got rid of the already_owned and already_added exceptions; they
  weren't being used anyway, and the above createAgent change makes them
  doubly unneeded.
- Made agents()/goods()/markets() return a const ref instead of a const
  copy of the maps involved.