Bertrand: allocation() fixes; buy() fixes
allocation() was sort of broken: it was just assigning the last unit
price instead of building up a total price from each firm.

allocation() was broken in that it assembled a quantity of 1 instead of
the requested quantity.

allocation() was really broken: the initial allocation struct wasn't
initialized!

Finished buy(): it now properly has the firms supply, transfers the
price from the buyer's assets to the firm's, then transfers the supplied
amount to the buyer's assets.