consumer/Compound.hpp: sums/products of consumers
Allowing arbitrary sums and products of consumers opens up lots of
possibilities without requiring a proliferation of consumer classes; the
new CompoundSum and CompoundProduct (and ::Differentiable variants)
allow just that.

As an example, that utility functions such as u(x,y,z) = 3 - xyz can be
represented as a CompoundSum of anything supporting a constant utility
and a CobbDouglas (with constant factor -1).