Bundle: yet another print fix
The previous commit fixed the problem for normal printing, but failed
for gtest's printing.  Added the Bundle-specific method back in, and
added a new protected _print method that does everything except print
the string "Bundle" or "BundleNegative"; the new ostream overrides now
just print either Bundle or BundleNegative, then call _print(os) to do
the rest of the printing.