Bundle.positive() returns positive part of Bundle
Sometimes its useful to treat a BundleNegative as if the negative
quantities don't exist (e.g. when the BN is some sort of assets
container).  Added positive() so that b.positive() returns a new Bundle
containing only the non-negative values of b.  (If b is already a
Bundle, there is a shortcut to just return a copy of b.)