Bundle.multiples: partial bundle division
I need something that does: 1/(b/a), which is *not* the same as a/b
bundle division.  Basically, a.multiples(b) tells me how many multiples
of b are contained within a, while a / b tells me how many b's I need to
meet or exceed a.

It turns out these are identical: 1/(b/a) and a.multiples(b), but the
latter is much clearer.