Market.quantity(p) changed to return more info
Rationale: quantity(p) sometimes doesn't spent all of p (because, for
instance, the market is constrained), but doesn't provide a nice way to
return that.  (You could get it by calling price() with the returned
quantity, but that's a bit cumbersome and inefficient).

Solution: quantity(p) now returns a struct indicating the quantity,
whether that quantity hit a constraint, and the actual spending (and
unspent money) that would result from the purchase.