interopt::Stepper: Added JumpStepper, periods
Stepper now takes an extra (optional) period argument, which allows
steps to be taken only every x periods.  Defaults to 1, which is
identical to the previous behaviour.

JumpStepper is a wrapper around Stepper that can take a large jump
(instead of a relative step), which is useful when initial conditions
are far away from equilibria (for instance, starting with quantity of
100 when equilibrium quantity is more like 5, or vice versa).