Position: at(d); algebra uses mutators; == & !=
pos.at(d) does exactly the same thing as pos[d], but is much more
convenient for internal code (at(i) instead of operator[](i) or
(*this)[i]).

Simplified code by making the algebra operators use the mutator versions
instead (i.e. a + b now copies a into a new object, then += b on it).
This also fixes some bugs because several algebra methods were
initializing the new position object to 0 instead of copying *this.

Added == and !=, for all-equal and not(all-equal).