Simulation: ...Filter caching
The voting simulation was running into a big performance bottleneck
because it needed to access all Party agents, of which there were 2
versus hundreds of Voter agents.  Each Filter call was thus terribly
slow, because it had to attempt to recast hundreds of objects to Party,
but only two would succeed.

This change introduces a filter cache so that when agentFilter<C> is
called with some agent subclass (but not the C=Agent) the objects that
match the class filter are stored so that next time we ask for
agentFilter<C>, we already have the list on hand.  The Agent-specific
part of the cache gets invalidated whenever a new Agent is added or an
existing one removed.  (Everything here applies in exactly the same way
to Good, Market, InterOptimizer, etc.).