MUPD: need to check feasibility *after* locking
MUPD was causing a feasibility error because market feasibility was
being checked before getting a market lock, which meant the market could
be feasible, but after the lock might not be feasible anymore, in which
case allocations failed with threads.  Moving the feasibility check to
after the big market write lock fixes that.