Member::Lock: added member removal
Added the ability to remove a subset of members from a Lock.  This
returns a new Lock consisting of the removed members--if the returned
lock isn't used, it is released immediately, thus freeing the lock on
the given members.

This also involved converting Lock's internal vector implementation to a
multiset, which required substantially reworking the locking code into
using iterators on the multiset instead of indices.