Fixed negativity error exceptions from MUPD change
Cancelling a reservation (which new MUPD code does sometimes) was not
quite working because the tolerance was too low and thus negativity
errors would result.  This was supposed to be handled in the Firm
already--and indeed it was within transfer(), which used transferApprox
with its own epsilon, but release() wasn't passing epsilon and was
getting Bundle's default (1e-13) instead of the desired 1e-10.