Removed interPostAdvance, added intraFinish
interopt::PostAdvance is an artifact from before the great optimizer
overhaul which no longer serves a purpose: got rid of it.  Coding using
interopt::PostAdvance should instead be using either interopt::Advance
(if the code was backwards-looking) or intraopt::Initialize (if the code
is doing things for the next period).

intraopt::Finish is a new optimizer for handling post-stage actions that
should run just before the stage is finished.  (For example:
transferring the period's money earnings from assets() into a profit
bundle).