Added weak dependencies.
Weak dependencies work like hard dependencies, except that instead of
triggering simulation removal when a dependency is removed it triggers
a (new) member.weakDepRemoved(removed_id) call.  This is particularly
useful for Markets, which have a list of firms, as the firms should be
removed from any Market it belongs to when being removed.  Handling
this elsewhere was pointlessly tedious: Market now declares a weak dep
on the firm when adding a firm, and removes the firm its list of firms
if the firm gets removed from the simulation.