Removed object cloning.  It doesn't work.
Cloning is not something that is particularly well tested or well
supported.  In particular, it currently fails entirely because Member
isn't copyable (due to having an uncopyable mutex member).  Making it
copyable, however, is probably a bad idea: copying a Market, for
example, which already has Firms is probably going to fail: the list of
constituent firms will be duplicated, but the Simulation's weak
references relationship won't be.

Rather than having a generic clone method that fails, better to not
have one (and let supporting classes make an explicit clone method if
they really want it).