Locking fixes for non-simulation members
Market reservations were causing weird errors in test code: whenever an
exception got thrown that resulted in a Reservation and market being
destroyed, a no_simulation_error() got triggered because the
reservation was trying to cancel itself, which led to something trying
to get a writeLock(), which in turn called maxThreads() which needs a
simulation().  Even without maxThreads() there's a problem, though,
because sharedSelf() won't work without a simulation, either.

Fixed this by making the locking code not try to lock when the caller
is not in a simulation.  This probably won't cause issues because it
seems only likely to happen when the caller is being destroyed or about
to be destroyed.

A related change: got rid of the zero-argument writeLock() and
readLock() methods from Member.cpp: they can be handled just as well
with the varargs version in Member.hpp, resulting in less code
duplication.  (There were three versions each of writeLock and
readLock(), now down to two).