Changed AssetAgent to no longer clear
There's currently no nice way to get an agent with assets without the
assets clearing at the end of each period other than overriding
interAdvance to not clear.

This commit fixes that: AssetAgent's interAdvance is now gone, moved to
AssetAgentClearing, which is a specialization of AssetAgent that adds
interAdvance asset clearing.  AssetAgent now *only* adds assets to the
base Agent class.