Added publically accessible runLock()
This lock can be created from outside the simulation so as to be able
to protected simulation access to in-between-stages (for example, for
updating a GUI showing current simulation status).  The lock is held
internally by run(), and so holding it guarantees run() isn't active
and won't proceed until the lock is released.