Made Member::Lock implement Lockable concept
Basically this required renaming release() to unlock() and adding a
try_lock().  The latter also carries with it new try options for read()
and write() (and _mutex_lock()), which now all return a bool to indicate
whether the lock succeeded (which, of course, is always true when
the try parameter is false, which is the default).