Added original type tracking.
This allows looser restrictions on the copy assignment operator: in
particular, it means you're now allowed to assign one SharedMember to
another *if* the *original* type of the two SharedMembers is the same.
In other words, if a and b are both SharedMember<Derived>, and Derived
inherits from Base, and am and bm are a and b cast into
SharedMember<Member>, then any of a = b, am = bm, am = b, and a = bm are
allowed (the latter two via an implicit cast constructor).

If c comes along which is originally a SharedMember<Derived2> (where
Derived2 is derived from Derived), then any attempt to assign between a
and c or b and c will fail, *regardless* of what SharedMember<T> c gets
cast into.  Even if SharedMember<Derived> cd{c} is created, cd is still
not interchangeable with a/am/b/bm because the original type differs.

In practice, this allows code such as:

    SharedMember<T> best;
    for (auto t : simulation()->agents<T>()) {
        if (!best.ptr() or t.better_than(best)) {
            best = t;
        }
    }

as long as T is a specific type (and not a base class that catches
other derived classes).

It also allow random shuffling via

    std::vector<SharedAgent<T>> v = ...;
    std::shuffle(v.begin(), v.end(), Random::rng());

(subject to the same caveat that everything in the vector must be an
instance of exactly the same class) which wasn't allowed before because
of the inability to copy.