Removed const ref args, optimize wrapping
Passing const ref for simple types is no faster (and can be slower).

Changed the wrapping data structure to a vector of bools with the value
at index i indicating whether dimension i wraps.  Previously it was a
set containing the indices, but since the number of indices is fixed and
known at construction, and because wrapped positional are likely to have
most (if not all) dimensions wrapped, using a simple vector address
should be considerably more efficient than the repeated binary search
lookups performed with the set.