Removed Eris<T> class
Eris<T> was a pointless wrapper and unnecessary wrapper around
std::shared_ptr<T>: this adds a public Simulation::spawn() static method
that creates a Simulation wrapped in a std::shared_ptr.  (spawn() was
used instead of create() because the latter is already taken for
constructing-and-adding a member).

This commit also makes the default constructor private instead of
protected (protected was already effectively private since Simulation is
declared final).

(This commit was also the reason for the version bump to 0.2.1 from a
couple of commits ago: that version change was supposed to be carried
in this commit).