Swapped spawn() and create() method names
Simulation::create() makes far more sense for creating a new simulation
object than Simulation::spawn(), and sim->spawn<T>(...) makes far more
sense for creating a new simulation member than sim->create<T>(...).

Since the new names don't interfere with the old ones (because the
add-a-member version has a required template argument, while
the create-a-simulation-object method has no templates), the old
versions can be left in place, and so are still there (simply
forwarding to the new version) and are now marked deprecated.

Bumped the version up to 0.2.3 to go along with this change.