Added repeated-boundary constructors
WrappedPositional<T> and Positional<T> can now be initialized with
two numeric boundaries, which are used as the boundaries in all
dimensions.  In other words:

WrappedPositional<T>({1,2,3}, 3, -3)

is now valid, and results in the same behaviour (i.e. wrapping in every
dimension) as:

WrappedPositional<T>({1,2,3}, {3,3,3}, {-3,-3,-3})

(and likewise for Positional<T>)