ERIS_DBG: try to print all at once
ERIS_DBG(a << b << c) called std::cerr << a << b << c, but that is three
separate print operations and so, when threading, it's quite common to
get abc not printed together (but instead intermixed with other thread
output).

This commit changes the debugging macros to stick the debugging output
into a std::ostringstream, then print that all at once.