Matrix interface; bump to 0.3.0 in CMakeLists
Added a matrix interface for use in Eris.  This interface is designed to
be generic rather than tied to any specific matrix implementation, and
only provides the functionality needed within eris; this commit includes
an Eigen implementation.  This eigen implementation is entirely
header-based, so that eris itself doesn't directly depend on Eigen.

v0.3.0 will incorporate this as well as the linear Bayesian belief
structure from creativity.