Position minor type fixes and subdimensions
Position indexing was via int instead of size_t for some unknown
reason.  This wasn't worth changing on its own, but since the following
already means an API change, might as well change it.

Position::subdimensions was particularly screwed up, typewise, as it
took *double* values for the indices.  Replaced it with a version that
takes size_t's instead, and added a generic container version that
takes any integer container.

Bumped build version to 0.3.2 and lib version.  (NB: this isn't tagged
yet; other changes may be included in the final 0.3.2 release).