Remove {Firm,Market}::Reservation misdirection
The nested Reservation classes in Firm and Market are convoluted due to
the use of a unique_ptr wrapper.

This replaces the whole mess with simpler classes that are move
constructable, which is all that was every really needed.

This also fixes the previous commit by moving the ReservationState
declaration to Firm.hpp, and unnesting it: it can't be in Market.hpp
because Firm needs it, but the #include dependency is Market -> Firm,
not Firm -> Market; and since both need it, it doesn't make sense to
nest it in one of them.