g++ 4.8 fix for forward-declarations
Under g++-4.8, the simAgent(), etc. calls changed here failed because
std::enable_if doesn't seem to work properly with forward-declared
classes (but the forward declarations are needed for the
interdependencies in eris).

This commit moves the calls into .cpp files (along with moving a couple
other code bits of code that didn't need to be in the .hpps).