BayesianLinear prediction fixes and enhancements
BayesianLinear prediction was pretty screwed up: the distribution of
predicted values was messed up, and error variances weren't being
incorporated at all.  This fixes it.

Since this is a pretty major fix, bumping the eris version.

This also changes the predict interface to be able to predict multiple
rows at once (instead of just one row), and adds a new predictGeneric()
method that allows prediction of arbitrary functions of y*.  One such
implementation -- which returns both the mean and variance of the
prediction values -- is included in the new predictVariance() method.

This, however, breaks API compatibility, so bumping the lib version.

linear-draw-test is a new scratch script that verifies that the
distributions are actually what Koop says they should be.

This also adds a (non-truncated) multivariate t generator to go along
with multivariateNormal(), and a rmvt scratch script to test it.