Fixed drawGibbs() distribution and draw failures
drawGibbs() was producing many draw failures, even in cases where
rejection sampling wasn't that bad (for example, I had a test case with
about a 20% rejection rate--enough to end up below the default 0.2
minimum success rate needed to keep using rejection sampling.

It appears that the code was drawing from the wrong distribution: where
it should be drawing independent normals with mean 0 and variance
sigma^2, it was actually drawing N(0,1)--which, when s^2 is large, puts
valid rejection ranges far out in the tails.  For example, with s^2 =
183, a value of -80 is about 5.9 standard deviations from 0 instead of
80 standard deviations from 0; the former can be drawn from, the latter
cannot (since the cdf values are smaller than DBL_MIN).