BayesianLinear update/weaken redesign
BayesianLinear.update() and .weaken() was sort of a mess, requiring
subclasses to import macros to have them return the right type.

The new design adds constructors for updating and/or weakening, so that
what used to be:

    BayesianLinear updated = prior.weaken(w).update(y, X);

becomes:

    BayesianLinear updated(prior, y, X, w);

This also allows for in-place updating/weakening via:

    model = BayesianLinear(std::move(prior), y, X, w);

which invokes a move constructor and move assignment, but both should be
fairly cheap.  This really couldn't be done before, because the only way
to get at the in-place methods would be via something like:

    model = std::move(model).weaken(w);

but that might not work because it essentially up invoking x =
std::move(x), which was not well-defined with the default move
constructor.

This breaks API compatibility, since the weaken/update methods are now
gone, so it requires an so version bump.  I'm also bumping the version
to 1.0.0 since the design is working fairly well and seems relatively
stable (there will be more changes in 1.0.0, so this commit is not
getting a version tag).