Returning sum was stupid: n isn't constant!
I was trying to figure out why I got different sums on different
compilers and at different optimization levels, but while the sum for a
given compiler/optimization was close on each run.  But of course n
depends on performance, so that's exactly what I should have expected.

This commit changes the code to use the draw means, not sums, and
results are as expected.