g++-4.9 fixes
g++ 4.9 doesn't appear to have std::defaultfloat (though it is a C++11
standard feature); work around it.

I also get warnings from the struct initialization with an empty set of
values, which *should* 0-initialize (and does, in g++ 5 and clang).  It
may also be 0 initializing in g++ 4.9, but is giving a warning, so this
fixes it.