Cost var overhaul; abstracted root; new a calcs
Instead of a smattering of doubles that needs new variables for each
library, stick everything in a map and loop through it to display
results.

Abstracted root-finding to a dedicated method, made a1 (which is
renamed from astar) use it.

Added aT and aTT1 to calculate critical a values at which:

   aT(n): For a values below this value, it is better to use the
   nth-order Taylor expansion of e^x (and accept the increase in
   expected cost) than to make the perfect decision using e^x.

   aTT1: For a values greater than this value, it is better to decide
   using an nth-order Taylor expansion rather than the (n-1)th order
   expansion.