Recalculate a0 without LambertW
The a0 calculation (sqrt(L) + 1/sqrt(L)), where L = LambertW(...) isn't
right, because it ignores the cost of the sqrt that is needed.  This
sqrt cost doesn't enter c_ER, because it isn't needed for an
Exponential rejection *iteration*, it can be calculated just once.

The solution to that equation means that the nice LambertW-using form is
gone, and so this code moves to solving for a0 by root solving (like a1,
aT1, etc.)--which is probably slightly more numerically accurate anyway,
since the LambertW calculation was only a numerical approximation (not
that it really matters for a0, which is based on random timings, and
thus probably only accurate to a couple decimal places at best anyway).