Added a new a0' variable for lambda=a approx
Geweke's implementation used exponential rejection with lambda=a
instead of lambda=0.5*(a + sqrt(a^2+4))--even though the latter has the
highest acceptance rate, Geweke writes that the increase isn't
worthwhile.

He's half right: there's a threshold value of a, which I'm calling a₀'
for now, above which it is better to use ER with lambda=a, and below
which calculation of the sqrt is worthwhile.

In essence, this means that for large a, Geweke's approach had the best
performance--it saves a sqrt on every draw, and for large a, lambda* is
very close to a anyway.