Revamped cost variable; pdf output; inverse mode
The costs of all the libraries is now in a nice list variable.

Output goes to a library-specific PDF file.

Output is produced for all libraries when run (one PDF per library).

Added an inverse mode for "fairytale" that produces the same
"acceptance rate" graphs as found in Li and Ghosh.  In theory, this can
be used for the other libraries too, but I'm not sure that this is
useful; does probability/mean(cost) have a useful interpretation?

vshift now shifts by device units, instead of user amounts.