Attempted fix for extra cycle on (d) overhead
I've been seeing, occassionally, an extra cycle for the (d) overhead
calculation.  This isn't consistent at all, though: rerunning (without
recompiling) can make it go away (or can make it appear, if it wasn't
there to start with).

This change updates the overhead to use 16-byte aligned, local
variables, which seems (based on a few runs without seeing the problem)
to fix things.