Another timing fix.
Timings were not quite right: some optimizations were getting through
that shouldn't be.  This commit should fix that by forcing the code to
evaluate a (known to be false) branch point in between the benchmark
call and the return mean accumulation.  This gives more consistent
results, and seems to be working well.