Tn reoptimizations; branch test; Tn@a test
The previous timing fix revealed some suboptimal Tn calculations; this
commit reoptimizes them.

Added a branch test to test that the compiler is doing proper
compile-time optimization of the constexpr-based branch.

Added timings for calculating a Taylor approximation of e^x with a !=
0.  This is basically slightly slower than the T{n+1} approximation (at
least for n=2 and n=3), and only achieves slight efficiency gains near
a0, while losing (more) efficiency at 0.  So not worth it, but worth
seeing the cost.