Incorporate HR/UR cost; one-sided short-circuiting
By not multiplying by the cost, the code was essentially treating cost
as 1 (instead of something more like 1/2), and so was using UR well into
the HR-is-better range.

This commit also adds checks for b == infinity, avoiding pointless
threshold calculations that will always (or never) be satisfied when b
is infinite (i.e. when doing one-sided truncation).