Better rejection performance calculations
Instead of trying to guess rejection calculation speed based on
individual operations, this change does it by actually doing a
feasible rejection iteration.  This gives considerably better results
for thresholds: speeds of the two methods are fairly close on either
side of a threshold, while before there were some large gaps.

This also adds use of the updated normal/exponential ziggurat
algorithms via the copy included in eris (calling it "boost+"), in
addition to the system boost version.

Updated the constants in truncated_normal_distribution.hpp resulting
from both of the above changes (mainly the first--the latter was
essentially already incorporated, because I was testing with boost git
with the updated algorithms).