Switch draw-perf to mt19937-64; use new normal/exp
eris is using mt19937_64, not mt19937 (because it seems to perform a
little better).  The draw perf results are actually a little weird:
normal and exponential get quite a bit faster, while uniform gets
slower.

many-randoms was still using the system boost normal/exponential
distributions, instead of the updated ones in eris::random.