Optimized decision code; updated thresholds/costs
This commit adds decision optimizations, reorganizing some of the
logic, eliminating some divisions, and being more efficient with the
cached variable calculations.

It also overhauls the threshold values, basing them on tested outcomes
rather than the draw-perf calculated values.

Finally this updates the relative costs based on the recent fix to
draw-perf (which was miscalculating ER cost because of the faulty
rejection probabilities); it also includes both the NR and HR costs,
instead of using the NR cost for both (HR is a little slower than NR).