draw failure resulted in uninitialized draw values
If a draw failure occurs (e.g. because of not enough admissable draws
when using rejection sampling), we'd end up with prediction_draws_
resized but not fully populated; a subsequent prediction attempt would
then use the uninitialized storage, which is bad (and was leading to
nonsensical things like a negative standard error).