Don't overallocate restriction matrix/vector
Just store the actual matrix/vector instead of something with rows >=
the number of restrictions.

This gets rid of the complication of needing a restrict_size_ with
matrix blocking all over the place.  This will incur a slight overhead
at construction (since the matrix needs to be resized for each new
restriction), but since the restriction matrix is generally tiny, but
we're multiplying by it multiple times when drawing, it seems like it
ought to be more efficient to be able to multiply by a full matrix
instead of a block of a matrix.  Performance aside, it also makes the
code simpler.