Exp acceptance rule was wrong for non-unit stdev
The x generated in the inner loop was a normalized value, but x_delta
was demeaned but not divided by sigma, which meant that (x-x_delta) was
just wrong (unless stdev = 1, of course).

This could be fixed by dividing x_scale by sigma, but that adds an
extra division, so instead this changes the inner-loop-generated x to be
a demeaned but not-normalized value (by multiplying it by _sigma), so
x-x_delta is a valid value.  This means, however, that (x-x_delta)^2
has been scaled by sigma^2, and so we need to multiply the LHS of the
acceptance condition by the same factor to get the right acceptance
rate.