New ER/UR decision logic; always use simple lambda
When ER is preferred to HR, always use lambda=a simplification: the
overhead of the full calculation seems so great that it is simply never
worthwhile.

When deciding between ER/UR, the new, simple threshold is:

    (b-a) < c/a

With c determined to be about 0.31 (tested on a couple Haswell
systems--on Sandy Bridge it seems to be slightly smaller at 0.29).  For
an actual exponential draw, this relationship holds exactly for any b >
a > 0.  For the exponential-rejection-for-normal, this relationship is
approximate, but is very, very close to exact beyond a of about 2 or 3
standard deviations, and isn't off by much even for much closer values
(for instance, at the current 0.55 HR/ER threshold, we should be using
UR for a slightly greater range (because UR does a bit better and ER
does a bit worse), but we're still within about 3% of optimal
performance.