Added program to calculate HR/UR crossover
The approach in the original (costless) paper, and even adapted with a
cost-based approach, won't work properly with the ziggurat normal draws:
the fundamental problem is that different draws take different amounds
of time, and the conditional normal/halfnormal draw cost is not a
constant.

This script tests it for a range of a and b-a values to attempt to
compute a linear approximation of the threshold, which seems to work
pretty well.