Use linear approximation of threshold line
This gets rid of the need for the Maclaurin approximation of exp
(instead of approximating just the exp for the theoretical performance,
now we're approximating the whole thing based on actual performance);
this also simplifies the HR-or-UR code, and should make the HR code
path a little faster (since it no longer needs to compute the 0.5/s^2
or a^2 values that UR needs, which were also, previously, needed in the
threshold calculation).

The coefficients (0.39 intercept and 0.42 slope) come from
random-hrur-crossover run on loki (Haswell).  The values on betwixt and
keynes are somewhat comparable (0.39 and 0.37 on betwixt (Sandy Bridge),
0.37 and 0.39 on keynes), all three using gcc 6.1.1.

On markov (also Haswell) with g++ 4.9, results are a bit different:
0.49 intercept, 0.33 slope, but I prefer to go with the optimum on the
newer CPU and compiler architecture as it seems more future-relevant,
and the other results, while a bit off, aren't wildly off.