Moved thresholds into detail; readd ER(lamba)
Moved the thresholds into a separate file.  The new values are
calculated by random-thresholds on loki, except for ur_below_nr_above,
which still needs to be implemented.

Getting rid of the ER(lamba*) values (done in
81e125c4c0ddfc79e6f8dc2ae49add34a701a9c2) was a mistake: there are some
big gains from using the full lambda calculation low in the distribution
(which makes sense, since the s in a + sqrt(a^2 + 4 s^2) is considerably
more important when a is small; e.g. for a=0.6, s=1 the proper lambda
is 1.34, which is not close to 0.6 at all.  At 1.33 (the new
threshold), the optimal value is 1.86, which is "only" 40% off versus
the 123% off at 0.6).