Correct truncated normal calculations for +-inf
Replace any terms of phi(x) times polynomial(x) with 0 when phi(x) == 0.

Since phi(a) is exponential, it goes to zero faster than any polynomial
goes to infinity, so all the terms involving phi(a) have a limit of 0
as a goes to infinity: do this, because otherwise the infinity screws
up the calculations.