Update threshold values
These are based on a 21 minute random-thresholds run on loki (i7-4770K).
Most don't change much; the biggest changes are the er(a)/er(l)
threshold, which is due to the fix in 15820cb; and the ur_below_nr_above
change from 1 to 0.969; the previous value of 1 was based on previous
guessing; 0.969 now comes from the threshold benchmarking.