many-randoms: accept runtime; use inline stat calc
Instead of just a number of draws, many-randoms now takes a time
argument (with an "s" suffix) to run for the given number of seconds.

Draws are now performed with a relatively small buffer size (16K
doubles) at a time, with inline mean/var/skew/kurt calculations instead
of storing all the draws in a big vector.

Output now includes mean ns/call (in addition to current Mdraws/s)