Added serialization base and helper classes
The base class handles compression and provides an interface for
reading/writing/updating a set of fixed parameter values and for
transparently handling compression.  Beyond that, it requires a subclass
to handle reading/writing the various serialized data.

Most of this design comes from creativity::state::FileStorage, which
will shortly be updated to use this new class.