enable_if via return type; removed depr. methods
enable_if was being done via extra template parameter in various
places, but it can be better done by applying to the return type,
making the templates cleaner.

This also bumps the version and soversion for a 0.5.0 pre-release.

Removed the deprecated create<T>(...) and spawn() methods: these
deprecated compatibility versions switched names in 0.2.3, nearly two
years ago; it seems an appropriate time to drop them.

Added simulation->add() method, which works like
simulation->spawn<T>(...) but takes a pre-created,
std::shared_ptr-wrapped member Member-derived object instead of
constructing one on the fly.  This is needed for the pyeris interface
since we need to construct member types dynamically.

Inlined short methods into the class definition.  The post-definition
versions got a little messier with the enable_if-on-return-type change,
so easier to just include the code with the method definition.